+++ test/new_relic/agent/utilization/aws_test.rb
@@ -65,6 +65,13 @@ module NewRelic
           refute_metrics_recorded "Supportability/utilization/aws/error"
         end
 
+        def test_that_the_headers_lambda_is_processed_properly
+          phony_imds_token = 'And with a token bird I made'
+          mock_response = mock(code: '200', body: phony_imds_token)
+          Net::HTTP.any_instance.stubs(:send_request).returns(mock_response)
+          assert_equal @vendor.headers, "X-aws-ec2-metadata-token" => phony_imds_token
+        end
+
         # ---
 
         def aws_fixture_path
@@ -77,7 +84,8 @@ module NewRelic
           test_case = symbolize_keys_in_object test_case
 
           define_method("test_#{test_case[:testname]}".gsub(" ", "_")) do
-            uri_obj = test_case[:uri][:'http://169.254.169.254/2016-09-02/dynamic/instance-identity/document']
+            NewRelic::Agent::Utilization::AWS.stubs(:imds_token).returns('John Howe')
+            uri_obj = test_case[:uri][:'http://169.254.169.254/latest/dynamic/instance-identity/document']
             if uri_obj[:timeout]
               @vendor.stubs(:request_metadata).returns(nil)
             else